I2C1和I2C2搞定,成功读写EEPROM以及0,91屏幕

+++ Core/Inc/i2c.h
@@ -68,12 +68,10 @@ void MX_I2C2_Init(void);
   /*I2C的普通传输:EEPROM:at24c256*/
 
 
-
   ErrorStatus my_I2C_At24c256_Write (I2C_TypeDef * I2Cx , uint8_t * p_data , uint32_t len , uint8_t slave_addr , uint16_t mem_addr);
 
 
-  ErrorStatus my_I2C_At24c256_Read (I2C_TypeDef * I2Cx , uint8_t * p_data , uint32_t len , uint8_t slave_addr , uint8_t mem_addr);
-
+  ErrorStatus my_I2C_At24c256_Read (I2C_TypeDef * I2Cx , uint8_t * p_data , uint32_t len , uint8_t slave_addr , uint16_t mem_addr);
